% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bdl_na.R
\name{any_bDL}
\alias{any_bDL}
\title{Looking for any "<" Detection Limit sign in a character vector}
\usage{
any_bDL(name)
}
\arguments{
\item{name}{character vector}
}
\description{
Looking for any "<" Detection Limit sign in a character vector
}
