% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dec_separator.R
\name{any_comma_decimal_separator}
\alias{any_comma_decimal_separator}
\title{Check if there is any comma used as decimal separator in a column vector}
\usage{
any_comma_decimal_separator(name)
}
\arguments{
\item{name}{string character}
}
\description{
Check if there is any comma used as decimal separator in a column vector
}
