% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bdl_na.R
\name{bDL}
\alias{bDL}
\title{Give a randomly selected value between 0 and the Detection Limit value}
\usage{
bDL(name, LoD = 0)
}
\arguments{
\item{name}{a character vector}

\item{LoD}{Limit Of Detection value}
}
\description{
Give a randomly selected value between 0 and the Detection Limit value
}
