% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boxplot.R
\name{boxplotServer}
\alias{boxplotServer}
\title{Server module rendering a boxplot. It can be linked with plotUi by id.}
\usage{
boxplotServer(id, data)
}
\arguments{
\item{id}{module identifier}

\item{data}{dataframe}
}
\description{
Server module rendering a boxplot. It can be linked with plotUi by id.
}
