% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boxplot.R
\name{boxplotUi}
\alias{boxplotUi}
\title{User Interface for boxplot with inputs and outputs}
\usage{
boxplotUi(id)
}
\arguments{
\item{id}{module identifier}
}
\description{
User Interface for boxplot with inputs and outputs
}
