% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{categoricDataset}
\alias{categoricDataset}
\title{extract a dataset with only categorical variables from a given dataset}
\usage{
categoricDataset(dataset)
}
\arguments{
\item{dataset}{a given dataframe}
}
\description{
extract a dataset with only categorical variables from a given dataset
}
