% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{categoricIndex}
\alias{categoricIndex}
\title{extract position indexes of the categorical variables in a given dataset}
\usage{
categoricIndex(dataset)
}
\arguments{
\item{dataset}{a given dataset}
}
\description{
extract position indexes of the categorical variables in a given dataset
}
