% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/varSelection.R
\name{catVarServer}
\alias{catVarServer}
\title{Server for catVarUi. Fills the catVarUi with the categorical variable names.}
\usage{
catVarServer(id, data)
}
\arguments{
\item{id}{module identifier}

\item{data}{data where the categorical variable names are found}
}
\description{
Server for catVarUi. Fills the catVarUi with the categorical variable names.
}
