% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/varSelection.R
\name{catVarUi}
\alias{catVarUi}
\title{SelectInput object with factor or character variables. It is the UI for catVarserver and they are linked by id.}
\usage{
catVarUi(id)
}
\arguments{
\item{id}{module identifier}
}
\description{
SelectInput object with factor or character variables. It is the UI for catVarserver and they are linked by id.
}
