% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{characterDataset}
\alias{characterDataset}
\title{extract a dataset with only character variables in a given dataset}
\usage{
characterDataset(dataset)
}
\arguments{
\item{dataset}{a given dataset}
}
\description{
extract a dataset with only character variables in a given dataset
}
