% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{characterIndex}
\alias{characterIndex}
\title{extract position indexes of the character variables in a given dataset}
\usage{
characterIndex(dataset)
}
\arguments{
\item{dataset}{a given dataset}
}
\description{
extract position indexes of the character variables in a given dataset
}
