% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classModels.R
\name{classModelUi}
\alias{classModelUi}
\title{Ui selectInput for some available classification models in CARET package}
\usage{
classModelUi(id)
}
\arguments{
\item{id}{module identifier}
}
\description{
Ui selectInput for some available classification models in CARET package
}
