% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classpred.R
\name{classPrediction}
\alias{classPrediction}
\title{Predict a sample of unseen data given fitted classification model}
\usage{
classPrediction(model, sample)
}
\arguments{
\item{model}{a given classification model}

\item{sample}{a sample to predict wit respect to the given classificatin model}
}
\description{
Predict a sample of unseen data given fitted classification model
}
