% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classpred.R
\name{classPredOutput}
\alias{classPredOutput}
\title{function server writing predictions into the classPredUi Ui}
\usage{
classPredOutput(id, pred)
}
\arguments{
\item{id}{module identifier}

\item{pred}{dataframe with predicted sample or list with predicted sample and unpredicted sample}
}
\description{
function server writing predictions into the classPredUi Ui
}
