% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classpred.R
\name{classPredServer}
\alias{classPredServer}
\title{compute the prediction of classes for the given model and sample in input}
\usage{
classPredServer(id, data1, data2)
}
\arguments{
\item{id}{module identifier}

\item{data1}{sample or model loaded in input}

\item{data2}{sample or model loaded in input}
}
\description{
compute the prediction of classes for the given model and sample in input
}
