% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classPredTest.R
\name{classPredTestServer}
\alias{classPredTestServer}
\title{Write the chosen classification model Test output into classPredTestServer Ui}
\usage{
classPredTestServer(id, data1, data2, pred)
}
\arguments{
\item{id}{module object identifier}

\item{data1}{data read from application}

\item{data2}{data read from application}

\item{pred}{predicted data from classPredServer}
}
\description{
Write the chosen classification model Test output into classPredTestServer Ui
}
