% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classPredTest.R
\name{classPredTestUi}
\alias{classPredTestUi}
\title{Ui for classification model test on unseen data}
\usage{
classPredTestUi(id)
}
\arguments{
\item{id}{module object identifier}
}
\value{
a shiny tagList object with different Ui for output visualization
}
\description{
Ui for classification model test on unseen data
}
