% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/granaManipulation.R
\name{convert_data_type}
\alias{convert_data_type}
\title{Convert data variables to the right data type}
\usage{
convert_data_type(data)
}
\arguments{
\item{data}{a dataframe}
}
\description{
Convert data variables to the right data type
}
