% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/granaManipulation.R
\name{data_preprocessing}
\alias{data_preprocessing}
\title{Compute some data preprocessing steps such as replace the under detection limit value wiht a random variable between 0 and the detecttion limit,
convert data variables and remove records with high number of missing values}
\usage{
data_preprocessing(data)
}
\arguments{
\item{data}{dataframe}
}
\description{
Compute some data preprocessing steps such as replace the under detection limit value wiht a random variable between 0 and the detecttion limit,
convert data variables and remove records with high number of missing values
}
