% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataViz.R
\name{dataVizOutput}
\alias{dataVizOutput}
\title{Show an overview of data in dataVizUi.}
\usage{
dataVizOutput(id, data)
}
\arguments{
\item{id}{module identifier}

\item{data}{data to overview for example output of the module created by fileUi and fileServer.}
}
\description{
Show an overview of data in dataVizUi.
}
