% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataViz.R
\name{dataVizUi}
\alias{dataVizUi}
\title{Ui where the data are overviewed by dataVizOutput function.}
\usage{
dataVizUi(id)
}
\arguments{
\item{id}{module identifier}
}
\description{
Ui where the data are overviewed by dataVizOutput function.
}
