% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.R
\name{downloadUi}
\alias{downloadUi}
\title{Ui for the downloadButton in the download module}
\usage{
downloadUi(id)
}
\arguments{
\item{id}{module identifier}
}
\description{
Ui for the downloadButton in the download module
}
