% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eda.R
\name{edaUi}
\alias{edaUi}
\title{User interface container for descriptive statistics output}
\usage{
edaUi(id)
}
\arguments{
\item{id}{identifier for the data module}
}
\description{
User interface container for descriptive statistics output
}
