% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{file_to_sqlite_dsn}
\alias{file_to_sqlite_dsn}
\title{Upload data file in sqlite database through available DSN.}
\usage{
file_to_sqlite_dsn(dsn, data, tablename, overwrite = FALSE, append = TRUE)
}
\arguments{
\item{dsn}{data source name}

\item{data}{data to be uploaded in the database}

\item{tablename}{data base table name}

\item{overwrite}{boolean value overwriting or not an existing table}

\item{append}{boolean value updating an existing table. If equals FALSE a new table is created}
}
\description{
Upload data file in sqlite database through available DSN.
}
