% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.R
\name{filedownServer}
\alias{filedownServer}
\title{Function server which downloads R objects load from datafile module (i.e fileUi & fileServer) when the downloadUi is clicked}
\usage{
filedownServer(id, data)
}
\arguments{
\item{id}{module identifier}

\item{data}{the R object to be saved in a binary file when the button is clicked}
}
\description{
Function server which downloads R objects load from datafile module (i.e fileUi & fileServer) when the downloadUi is clicked
}
