% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bdl_na.R
\name{fill_bDL}
\alias{fill_bDL}
\title{Fill the below detection limit missing value with a random number between 0 and the the detection limit threshold}
\usage{
fill_bDL(data)
}
\arguments{
\item{data}{dataframe}
}
\description{
Fill the below detection limit missing value with a random number between 0 and the the detection limit threshold
}
