% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classModels.R
\name{fit_class}
\alias{fit_class}
\title{Fit and train some classification models}
\usage{
fit_class(data, catVar, model)
}
\arguments{
\item{data}{a dataset with explanatory variables and response variable}

\item{catVar}{the response variable}

\item{model}{string specifiying which classification model will be used.}
}
\value{
a trained and fitted classification model
}
\description{
Fit and train some classification models
}
