% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bdl_na.R
\name{impute_bDL}
\alias{impute_bDL}
\title{fill <DL position in a dataset with a specific value given randomly between 0 and the LoD}
\usage{
impute_bDL(data)
}
\arguments{
\item{data}{a dataframe}
}
\description{
fill <DL position in a dataset with a specific value given randomly between 0 and the LoD
}
