% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bdl_na.R
\name{is_bDL}
\alias{is_bDL}
\title{Looking for the below Detection Limit sign "<" into a character vector or a dataframe}
\usage{
is_bDL(name)
}
\arguments{
\item{name}{dataframe name or variable name of character vector}
}
\value{
a boolean vector or a dataframe of boolean values
}
\description{
Looking for the below Detection Limit sign "<" into a character vector or a dataframe
}
