% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dec_separator.R
\name{is_point_decimal_separator}
\alias{is_point_decimal_separator}
\title{Check if the point is the decimal separator}
\usage{
is_point_decimal_separator(name)
}
\arguments{
\item{name}{string character}
}
\description{
Check if the point is the decimal separator
}
