% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parcoord.R, R/plot.R
\name{multiplotOutput}
\alias{multiplotOutput}
\title{Server module rendering a parallel plot coordinate. It can be linked with parcoordUi by id.}
\usage{
multiplotOutput(id, data)

multiplotOutput(id, data)
}
\arguments{
\item{id}{identifier of the module}

\item{data}{dataframe}
}
\description{
Server module rendering a parallel plot coordinate. It can be linked with parcoordUi by id.

Server module rendering multiple plot containers for each level of a categoricalvariable.
If two or more categorical variable are selected, the first one will be chosen.
}
