% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bdl_na.R
\name{n_bDL}
\alias{n_bDL}
\title{Count the number of values below the limit of detection}
\usage{
n_bDL(name)
}
\arguments{
\item{name}{variable name of the character vector}
}
\description{
Count the number of values below the limit of detection
}
