% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bdl_na.R
\name{n_missing}
\alias{n_missing}
\title{Count the number of missing values in a vector}
\usage{
n_missing(name)
}
\arguments{
\item{name}{variable name of the character vector}
}
\description{
Count the number of missing values in a vector
}
