% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bdl_na.R
\name{n_missing_bDL}
\alias{n_missing_bDL}
\title{Count the number of missing values and/or values below the detection limit in a vector}
\usage{
n_missing_bDL(name)
}
\arguments{
\item{name}{variable name of the character vector}
}
\description{
Count the number of missing values and/or values below the detection limit in a vector
}
