% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/notifications.R
\name{notificationServer}
\alias{notificationServer}
\title{Server function throwing error or warning to notificationUi.}
\usage{
notificationServer(id, expr)
}
\arguments{
\item{id}{module identifier}

\item{expr}{expression to be evaluated}
}
\description{
Server function throwing error or warning to notificationUi.
}
