% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/notifications.R
\name{notificationUi}
\alias{notificationUi}
\title{Ui for error or warning notifications}
\usage{
notificationUi(id)
}
\arguments{
\item{id}{module identifier}
}
\description{
Ui for error or warning notifications
}
