% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{numericDataset}
\alias{numericDataset}
\title{extract a dataset with only numeric variables from a given dataset}
\usage{
numericDataset(dataset)
}
\arguments{
\item{dataset}{a dataframe}
}
\description{
extract a dataset with only numeric variables from a given dataset
}
