% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{numericIndex}
\alias{numericIndex}
\title{extract indexes of the numeric variables from a given dataset}
\usage{
numericIndex(dataset)
}
\arguments{
\item{dataset}{a dataframe}
}
\description{
extract indexes of the numeric variables from a given dataset
}
