% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/varSelection.R
\name{numVarServer}
\alias{numVarServer}
\title{Fills the numVarUi with numeric variables names found in given data. It is linked with numVarUi by id}
\usage{
numVarServer(id, data)
}
\arguments{
\item{id}{module identifier}

\item{data}{data where the numeric variables names are found.}
}
\description{
Fills the numVarUi with numeric variables names found in given data. It is linked with numVarUi by id
}
