% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/varSelection.R
\name{numVarUi}
\alias{numVarUi}
\title{selectInput object for numeric variables. It is the UI for numVarServer and they are linked by id}
\usage{
numVarUi(id)
}
\arguments{
\item{id}{module identifier}
}
\description{
selectInput object for numeric variables. It is the UI for numVarServer and they are linked by id
}
