% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pca.R
\name{pcaOutput}
\alias{pcaOutput}
\title{output summary of pca}
\usage{
pcaOutput(id, data, pcaObject)
}
\arguments{
\item{id}{identifier of data module}

\item{data}{data used to fit the pca. It only necessary for biplot in order to classify the graphical visualization of scores.}

\item{pcaObject}{a given fitted pca model}
}
\description{
output summary of pca
}
