% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pca.R
\name{pcaServer}
\alias{pcaServer}
\title{Compute Principal Component Analysis with variable selection inputs and outputs summary with plots}
\usage{
pcaServer(id, data)
}
\arguments{
\item{id}{identifier of data module}

\item{data}{a given dataframe whose numeric variables will be extracted}
}
\value{
a pca object
}
\description{
Compute Principal Component Analysis with variable selection inputs and outputs summary with plots
}
