% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pca.R
\name{pcaUi}
\alias{pcaUi}
\title{User interface where the Principal Component Analysis outputs will be written with variable selection inputs}
\usage{
pcaUi(id)
}
\arguments{
\item{id}{data module identifier}
}
\value{
a tagList object
}
\description{
User interface where the Principal Component Analysis outputs will be written with variable selection inputs
}
