% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dec_separator.R
\name{replace_decimal_separator}
\alias{replace_decimal_separator}
\title{Replace the comma used as decimal separator with the point}
\usage{
replace_decimal_separator(data)
}
\arguments{
\item{data}{dataframe}
}
\description{
Replace the comma used as decimal separator with the point
}
