% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report.R
\name{reportServer}
\alias{reportServer}
\title{Server function for reportUi in charge of generate report}
\usage{
reportServer(id, report_script)
}
\arguments{
\item{id}{module identifier}

\item{report_script}{file with extension .Rmd in charge of parameterized report}
}
\description{
Server function for reportUi in charge of generate report
}
