% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bdl_na.R
\name{rm_bDL}
\alias{rm_bDL}
\title{Delete samples with high percentage of varialbes below de Detection Limit.
The threshold is given by more than 50\% of the numeric predictors.}
\usage{
rm_bDL(data)
}
\arguments{
\item{data}{dataframe}
}
\description{
Delete samples with high percentage of varialbes below de Detection Limit.
The threshold is given by more than 50\% of the numeric predictors.
}
