% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pca.R
\name{RobustPCAUi}
\alias{RobustPCAUi}
\title{Ui for outlier detection with robust pca plots}
\usage{
RobustPCAUi(id)
}
\arguments{
\item{id}{module object identifier}
}
\value{
a tagList shiny object
}
\description{
Ui for outlier detection with robust pca plots
}
