% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{sheet_list}
\alias{sheet_list}
\title{give the list of sheets in a workbook excel or tables in a microsoft access or SQLite database}
\usage{
sheet_list(name, path)
}
\arguments{
\item{name}{file name with extension}

\item{path}{the full path of the file}
}
\value{
a character vector of table names or sheet names
}
\description{
give the list of sheets in a workbook excel or tables in a microsoft access or SQLite database
}
