% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sqliteDSN.R
\name{sqliteDSNServer}
\alias{sqliteDSNServer}
\title{Function server for sqliteDSNUi. Read and load sqlite database table list through DSN}
\usage{
sqliteDSNServer(id)
}
\arguments{
\item{id}{module identifier}
}
\description{
Function server for sqliteDSNUi. Read and load sqlite database table list through DSN
}
