% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sqliteDSN.R
\name{sqliteDSNUi}
\alias{sqliteDSNUi}
\title{Ui reading DSN with available database connection}
\usage{
sqliteDSNUi(id)
}
\arguments{
\item{id}{module identifier}
}
\description{
Ui reading DSN with available database connection
}
