% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SQLiteFile.R
\name{SQLiteFileUi}
\alias{SQLiteFileUi}
\title{Ui for file give to SQLiteServer function.}
\usage{
SQLiteFileUi(id)
}
\arguments{
\item{id}{module identifier}
}
\description{
Ui for file give to SQLiteServer function.
}
