% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statsReport.R
\name{statsReportUi}
\alias{statsReportUi}
\title{User interface where for descriptive statistics}
\usage{
statsReportUi(id)
}
\arguments{
\item{id}{identifier for the data module}
}
\value{
a tagList object
}
\description{
User interface where for descriptive statistics
}
