% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/training.R
\name{trainingOutput}
\alias{trainingOutput}
\title{Server function writing the fitted classifier model into the rfTrainingUi}
\usage{
trainingOutput(id, object)
}
\arguments{
\item{id}{module identifier}

\item{object}{a fitted model object given by knnServer, rfServer etc....}
}
\description{
Server function writing the fitted classifier model into the rfTrainingUi
}
