% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/training.R
\name{trainingUi}
\alias{trainingUi}
\title{Ui container for the fitted classifier model}
\usage{
trainingUi(id)
}
\arguments{
\item{id}{module object identifier}
}
\description{
Ui container for the fitted classifier model
}
