% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classModelBuilding.R
\name{TrainTestModelOutput}
\alias{TrainTestModelOutput}
\title{Write the results outputs given by TrainTestModelServer in TrainTestModelUi}
\usage{
TrainTestModelOutput(id, data)
}
\arguments{
\item{id}{module identifier}

\item{data}{list of objects given by TrainTestModelServer containing the fitted model from training and the predictions from test}
}
\description{
Write the results outputs given by TrainTestModelServer in TrainTestModelUi
}
