% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classModelBuilding.R
\name{TrainTestModelServer}
\alias{TrainTestModelServer}
\title{Train and Test classification models such as RF, GBM, XGBTREE}
\usage{
TrainTestModelServer(id, data)
}
\arguments{
\item{id}{module identifier}

\item{data}{training data used by models}
}
\value{
a list of two objects: the fitted model, and the predictions on test sample
}
\description{
Train and Test classification models such as RF, GBM, XGBTREE
}
