% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classModelBuilding.R
\name{TrainTestModelUi}
\alias{TrainTestModelUi}
\title{Ui container for classification model Training and Test}
\usage{
TrainTestModelUi(id)
}
\arguments{
\item{id}{module identifier.}
}
\description{
Ui container for classification model Training and Test
}
