% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/varSelection.R
\name{varSelectionServer}
\alias{varSelectionServer}
\title{Server module for varPlotUi in charge of loading the selectInput with the corresponding variables}
\usage{
varSelectionServer(id, data)
}
\arguments{
\item{id}{module identifier}

\item{data}{dataframe for variable selection}
}
\description{
Server module for varPlotUi in charge of loading the selectInput with the corresponding variables
}
