% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/varSelection.R
\name{varSelectionUi}
\alias{varSelectionUi}
\title{Input module for variable selection in a dataset}
\usage{
varSelectionUi(id)
}
\arguments{
\item{id}{identifier of module object}
}
\value{
a taglist with 2 selectInput for categorical and numerical variables, and the output container for the plot
}
\description{
Input module for variable selection in a dataset
}
