% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SQLiteFile.R
\name{write_to_sqlite_file}
\alias{write_to_sqlite_file}
\title{Write to SQLite Database table through local SQLite Database File}
\usage{
write_to_sqlite_file(dbname, tbname, data, overwrite = FALSE, append = TRUE)
}
\arguments{
\item{dbname}{database name}

\item{tbname}{name for the created table}

\item{data}{data}

\item{overwrite}{boolean value whether the table is overwritten}

\item{append}{boolean value wheter the table is updated if TRUE or created if FALSE}
}
\description{
Write to SQLite Database table through local SQLite Database File
}
