% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upload.R
\name{writeToDBUi}
\alias{writeToDBUi}
\title{Ui for actionButton which writes into the database from an uploaded raw data file}
\usage{
writeToDBUi(id)
}
\arguments{
\item{id}{module identifier}
}
\description{
Ui for actionButton which writes into the database from an uploaded raw data file
}
