% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upload.R
\name{writeToSQLiteDSN}
\alias{writeToSQLiteDSN}
\title{Function server in order to write into sqlite database when the writeToDBUi button is clicked.
Overwrite a table is not considered, we only implement updating and creating table.}
\usage{
writeToSQLiteDSN(id, dat)
}
\arguments{
\item{id}{module identifier}

\item{dat}{data to be written into the database}
}
\description{
Function server in order to write into sqlite database when the writeToDBUi button is clicked.
Overwrite a table is not considered, we only implement updating and creating table.
}
