% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SQLiteFile.R
\name{writeToSQLiteFile}
\alias{writeToSQLiteFile}
\title{Function server in order to write into sqlite database file when the writeToDBUi button is clicked.
Overwrite a table is not considered, we only implements updating and creating table.}
\usage{
writeToSQLiteFile(id, dat)
}
\arguments{
\item{id}{module identifier}

\item{dat}{data to be written into the database}
}
\description{
Function server in order to write into sqlite database file when the writeToDBUi button is clicked.
Overwrite a table is not considered, we only implements updating and creating table.
}
