% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dec_separator.R
\name{wrong_decimal_separator_data}
\alias{wrong_decimal_separator_data}
\title{Detect columns with comma as decimal decimal separator}
\usage{
wrong_decimal_separator_data(data)
}
\arguments{
\item{data}{dataframe}
}
\description{
Detect columns with comma as decimal decimal separator
}
